+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.0] 2025-05-23
+### Added
+- Added `iframeChatTitle` config key for iframe `title` selector targeting
+- Added `iframeBackgroundColor` config key to prevent underlying elements showing through
+
+### Fixed
+- Chat iframe no longer overlaps the safe area
+
 ## [1.1.0] 2023-08-05
 ### Added
 - Support for PWA cookie consent feature. Widget initialization is postponed till users finished the consent process. When no "comfort" consent is given, the widget will not show up.